Temporarily disabled cleanup phase of build process, so I can point Firefox/Thunderbird directly to the build directory, rather than a prepackaged extension Removed useless uo->ou rule; added uo->ươ REP

+++ firefox_thunderbird/build.sh
@@ -147,8 +147,13 @@ else
   rm ./files
 fi
 
-# remove the working files
-rm -rf $TMP_DIR
+if [ $CLEAN_UP = 0 ]; then
+  echo
+else
+  echo "Cleanup..."
+  # remove the working files
+  rm -rf $TMP_DIR
+fi
 echo "Done!"
 
 $AFTER_BUILD
